+++ backend/database/logging/database.js
@@ -1,6 +1,7 @@
 const sqlite3 = require('sqlite3').verbose();
 const crypto = require('crypto');
-const db = new sqlite3.Database('data/logging_database.db', (err) => {
+const path = require('path');
+const db = new sqlite3.Database(path.resolve(__dirname, '../../storage/logging_database.db'), (err) => {
   if (err) {
     console.error('Error opening database:', err.message);
     throw err;
