+++ webscraping/cron.js
@@ -1,14 +1,14 @@
 const { scraperManager } = require('./scraperManager.js');
 const { scheduleJob } = require('node-schedule');
 const dotenv = require('dotenv');
-const { dbAll } = require('../database/database.js');
-const { checkAndRenewCredits } = require('../database/credits.js');
-const { getCurrentDate } = require('../database/dateUtils.js');
+const { dbAll } = require('../backend/database/database.js');
+const { checkAndRenewCredits } = require('../backend/database/credits.js');
+const { getCurrentDate } = require('../backend/database/dateUtils.js');
 dotenv.config();
 
 const reCrawlInterval = process.env.RE_CRAWL_INTERVAL || 1440; // 1440 minutes = 24 hours
 const cronCheckInterval = process.env.CRON_CHECK_INTERVAL || 15; // 15 minutes in production, 2 minutes in development
-const { getWebsitesByLastScrapedBefore } = require('../database/websites');
+const { getWebsitesByLastScrapedBefore } = require('../backend/database/websites.js');
 let cronsScheduled = false;
 
 async function checkRenewalOnCredits() {
