+++ website-api/conversations.js
@@ -1,8 +1,8 @@
 const express = require('express');
 const router = express.Router();
-const { storeConversation, getConversationsByChatbot, deleteConversation, getConversationById } = require('../database/conversations');
+const { storeConversation, getConversationsByChatbot, deleteConversation, getConversationById } = require('../backend/database/conversations');
 const { authMiddleware } = require('./middleware');
-const { getChatbot } = require('../database/chatbots');
+const { getChatbot } = require('../backend/database/chatbots');
 
 // Get conversations for a chatbot with pagination and filters
 router.get('/:chatbotId', authMiddleware, async (req, res) => {
