+++ website-api/files.js
@@ -4,8 +4,8 @@ const multer = require('multer');
 const path = require('path');
 const mime = require('mime-types');
 const { authMiddleware } = require('./middleware');
-const { addFile, updateFileVisibility, updateFileReferencing, updateFileTextContent, deleteFile, uploadsDir, getFilesByWebsiteId } = require('../database/files');
-const { getChatbotFromPlanId } = require('../database/chatbots');
+const { addFile, updateFileVisibility, updateFileReferencing, updateFileTextContent, deleteFile, uploadsDir, getFilesByWebsiteId } = require('../backend/database/files');
+const { getChatbotFromPlanId } = require('../backend/database/chatbots');
 const TextExtractor = require('./textExtractor');
 const { logger } = require('../utils/fileLogger');
 // Configure multer for file uploads
